+++ engine/open.go
@@ -224,6 +224,7 @@ func Open(path string, opts ...Option) (*Store, error) {
 	s.Query = &Query{store: s}
 	s.WorkflowStates = &WorkflowStates{store: s}
 	s.Labels = &Labels{store: s}
+	s.Documents = &Documents{store: s}
 
 	return s, nil
 }